+++ src/store/apis/archiver-v2.api.ts
@@ -1,8 +1,12 @@
 import { envConfig } from '@app/configs'
 import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react'
 import type {
+  GetEpochTicksArgs,
+  GetEpochTicksResponse,
   GetIdentityTransfersArgs,
   GetIdentityTransfersResponse,
+  GetTickTransactionsArgs,
+  GetTickTransactionsResponse,
   GetTransactionResponse
 } from './archiver-v2.types'
 
@@ -23,6 +27,18 @@ export const archiverV2Api = createApi({
         `identities/${addressId}/transfers?startTick=${startTick}&endTick=${endTick}`,
       transformResponse: (response: GetIdentityTransfersResponse) =>
         response.transactions.flatMap(({ transactions }) => transactions)
+    }),
+    getTickTransactions: builder.query<
+      GetTickTransactionsResponse['transactions'],
+      GetTickTransactionsArgs
+    >({
+      query: ({ tick, transfers = false, approved = false }) =>
+        `ticks/${tick}/transactions?transfers=${transfers}&approved=${approved}`,
+      transformResponse: (response: GetTickTransactionsResponse) => response.transactions
+    }),
+    getEpochTicks: builder.query<GetEpochTicksResponse, GetEpochTicksArgs>({
+      query: ({ epoch, pageSize, page }) =>
+        `epochs/${epoch}/ticks?desc=true&pageSize=${pageSize}&page=${page}`
     })
   })
 })
@@ -30,5 +46,7 @@ export const archiverV2Api = createApi({
 export const {
   useGetTransactionQuery,
   useGetIndentityTransfersQuery,
-  useLazyGetIndentityTransfersQuery
+  useLazyGetIndentityTransfersQuery,
+  useGetTickTransactionsQuery,
+  useGetEpochTicksQuery
 } = archiverV2Api