+++ src/store/index.ts
@@ -1,4 +1,5 @@
 import { configureStore } from '@reduxjs/toolkit'
+import { archiverV1Api } from './apis/archiver-v1.api'
 import { archiverV2Api } from './apis/archiver-v2.api'
 import localeReducer from './localeSlice'
 import { networkReducer } from './network/networkReducer'
@@ -9,10 +10,12 @@ export const store = configureStore({
     locale: localeReducer,
     search: searchReducer,
     network: networkReducer,
+    [archiverV1Api.reducerPath]: archiverV1Api.reducer,
     [archiverV2Api.reducerPath]: archiverV2Api.reducer
   },
 
-  middleware: (getDefaultMiddleware) => getDefaultMiddleware().concat(archiverV2Api.middleware)
+  middleware: (getDefaultMiddleware) =>
+    getDefaultMiddleware().concat(archiverV1Api.middleware).concat(archiverV2Api.middleware)
 })
 
 export type RootState = ReturnType<typeof store.getState>